backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
google-genai>=0.1.0
python-dotenv>=1.0.0
aiohttp>=3.10.0
Pillow>=11.0.0
imagehash>=4.3.1
elevenlabs>=1.0.0
PyMuPDF>=1.24.0
tripo3d>=0.1.0
rembg[cpu]>=2.0.50
